Translation and Flag of PT-BR
In the setup-zero.js file, the Brazil flag was added to select the PT-BR translation.
A Scroll has been added to allow the addition of new translations.

In the setup-three.js file. You can now save the name by clicking the keyboard or by the button (as before)

Translated the fixed text of 'Choose One', to dynamic according to the selected language. in the Announcer.js file.